Nope, Revert. FAdmin stores values as 'string' so it can be float

+++ gamemode/modules/fadmin/fadmin/serversettings/cl_init.lua
@@ -21,7 +21,7 @@ local function SetLimits()
 			left:SetValue(GetConVar(v.name):GetFloat())
 
 			function left:OnValueChanged(val)
-				if val == GetConVar(v.name):GetInt() then
+				if val == GetConVar(v.name):GetFloat() then
 					return
 				end
 				RunConsoleCommand("_FAdmin", "ServerSetting", v.name, val)
